Fix case where CFBundleDisplayName is null
Based on [Apple's
documentation](https://developer.apple.com/library/ios/documentation/Gen
eral/Reference/InfoPlistKeyReference/Articles/CoreFoundationKeys.html#//
apple_ref/doc/plist/info/CFBundleDisplayName), CFBundleDisplayName is
optional, so it's not safe for one to assume it's defined. As a
fallback we can check CFBundleName, which is guaranteed to be defined.
This fixes the issue for me.